Разнес функции по отдельным файлам и переименовал их, чтобы все не было в куче
